little mess with Mutating headers
