+++ quesma/v2/core/quesma_apis.go
@@ -28,6 +28,7 @@ type HTTPFrontendConnector interface {
 	FrontendConnector
 	AddRouter(router Router)
 	GetRouter() Router
+	MutateResponseWriter(w http.ResponseWriter) http.ResponseWriter
 }
 
 type TCPFrontendConnector interface {
